refactor: move refresh-all scheduler logic to task module
