+++ server/workers/update-scheduler.ts
@@ -7,8 +7,7 @@ import { db } from "../utils/db"
 import { findFingerprintPosition } from "../utils/fingerprint"
 import { getSourceById } from "../utils/sources"
 import type { SourceProvider } from "../utils/sources/core"
-import { calculateStaggerIntervalMs, isEligibleForRefresh } from "../utils/workers/update-scheduler-policy"
-import { handleRecomputeAllTask, handleRetryFailedPagesTask } from "../utils/workers/update-scheduler-tasks"
+import { handleRecomputeAllTask, handleRefreshAllTask, handleRetryFailedPagesTask } from "../utils/workers/update-scheduler-tasks"
 
 /**
  * FETCH_LATEST task: Check latest updates from each source and queue matching series.
@@ -158,84 +157,6 @@ async function handleFetchLatest(
 	await job.updateProgress(100)
 }
 
-/**
- * REFRESH_ALL task: Queue all tracked series with staggered delays.
- */
-async function handleRefreshAll(job: Job<UpdateSchedulerJobData>) {
-	const config = useRuntimeConfig()
-	const SPREAD_MS = config.schedulerRefreshSpreadMs
-
-	job.log("Starting REFRESH_ALL task")
-	await job.updateProgress(5)
-
-	const dbSources = await db.source.findMany({
-		where: { enabled: true },
-		select: {
-			id: true,
-			external_id: true,
-			rate_limit_max: true,
-			rate_limit_duration: true,
-		},
-	})
-
-	job.log(`Found ${dbSources.length} sources`)
-	let totalQueued = 0
-
-	for (const dbSource of dbSources) {
-		job.log(`Processing source: ${dbSource.external_id}`)
-
-		// Get tracked SerieSources for this source with their scheduling info
-		const trackedSerieSources = await db.serieSource.findMany({
-			where: { source_id: dbSource.id },
-			select: {
-				id: true,
-				external_id: true,
-				last_checked_at: true,
-				consecutive_failures: true,
-			},
-		})
-
-		job.log(`Source ${dbSource.external_id} has ${trackedSerieSources.length} tracked series`)
-
-		const now = new Date()
-		const seriesToRefresh = trackedSerieSources.filter((serieSource) => {
-			return isEligibleForRefresh({
-				lastCheckedAt: serieSource.last_checked_at,
-				consecutiveFailures: serieSource.consecutive_failures,
-				now,
-			})
-		})
-
-		job.log(`${seriesToRefresh.length} series pass backoff filter`)
-
-		const actualInterval = calculateStaggerIntervalMs({
-			rateLimitDurationMs: dbSource.rate_limit_duration,
-			rateLimitMax: dbSource.rate_limit_max,
-			spreadMs: SPREAD_MS,
-			totalItems: seriesToRefresh.length,
-		})
-
-		job.log(`Stagger interval: ${Math.round(actualInterval / 1000)}s between updates`)
-
-		for (const [i, serieSource] of seriesToRefresh.entries()) {
-			const delay = i * actualInterval
-
-			await serieInserterQueue.add(
-				"serie-inserter",
-				{ source_id: dbSource.id, source_serie_id: serieSource.external_id },
-				{ delay, priority: JOB_PRIORITY.NORMAL },
-			)
-
-			totalQueued++
-		}
-
-		job.log(`Queued ${seriesToRefresh.length} series from ${dbSource.external_id}`)
-	}
-
-	job.log(`REFRESH_ALL complete. Total queued: ${totalQueued}`)
-	await job.updateProgress(100)
-}
-
 export default defineWorker<typeof QUEUE_NAME, UpdateSchedulerJobData, undefined>({
 	name: QUEUE_NAME,
 	options: {
@@ -249,7 +170,7 @@ export default defineWorker<typeof QUEUE_NAME, UpdateSchedulerJobData, undefined
 			await handleFetchLatest(job, data.sourceId)
 		}
 		else if (data.type === "REFRESH_ALL") {
-			await handleRefreshAll(job)
+			await handleRefreshAllTask(job)
 		}
 		else if (data.type === "RETRY_FAILED_PAGES") {
 			await handleRetryFailedPagesTask(job)
